added email templates so that we can now send mails directly
