+++ one_big_thing/templates/email/provide-feedback.txt
@@ -1,0 +1,17 @@
+# We've noticed you have recorded 7 hours or more on the One Big Thing Platform but have not provided feedback.
+
+We hope you are enjoying your data training!
+
+Your feedback is essential to helping us evaluate the impact of One Big Thing 2023 and is your opportunity to let us know what you thought of the training. This will help us to improve future years of the initiative.
+
+To do this, [access](https://onebigthing.civilservice.gov.uk?utm_source=NudgeEmail&utm_medium=Email&utm_campaign=DecemberFeedback&utm_content=ProvideFeedbackLink)  the platform and the then scroll down to the feedback section on the homepage.
+
+We do not share any personal data that can identify you. Your answers will be aggregated and used for anonymised analysis. To understand more about how we will use your personal data, you can read the full privacy notice, available on the Platform.
+
+If you explicitly consent in the feedback survey to be contacted further about One Big Thing, researchers/analysts within the Cabinet Office may reach out to you for further information.
+
+Many thanks,
+One Big Thing Team
+Part of the [Modernisation and Reform Unit](https://www.gov.uk/government/organisations/modernisation-and-reform)
+
+^You are receiving this email because you have signed up to the One Big Thing Platform. 'One Big Thing' is a new annual initiative for civil servants to take shared action around a reform priority, as one Civil Service. The 'One Big Thing' for 2023 is data - all civil servants are being asked to complete at least one day's worth of data training this autumn. Have you provided feedback on your One Big Thing experience yet?
