+++ one_big_thing/templates/email/recorded-your-learning.txt
@@ -1,0 +1,13 @@
+# We've noticed you have signed up to the One Big Thing Platform but not recorded your 7 hours of data learning yet.
+
+To do this, [access](https://onebigthing.civilservice.gov.uk?utm_source=NudgeEmail&utm_medium=Email&utm_campaign=December&utm_content=RecordLearningLink)  the platform and head over to 'Record Learning' on the top navigation bar.
+
+It is important to record your learning as it helps us evaluate and assess the value of One Big Thing and improve future years of this initiative. You can also email your learning record to yourself, your manager or anyone else.
+
+Please note that this platform is separate to Civil Service Learning and any departmental learning platform. You may have recorded learning there but please also record it on the OBT Platform.
+
+Many thanks,
+One Big Thing Team
+Part of the [Modernisation and Reform Unit](https://www.gov.uk/government/organisations/modernisation-and-reform)
+
+^You are receiving this email because you have signed up to the One Big Thing Platform. 'One Big Thing' is a new annual initiative for civil servants to take shared action around a reform priority, as one Civil Service. The 'One Big Thing' for 2023 is data - all civil servants are being asked to complete at least one day's worth of data training this autumn.
